hackerstreet.6/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
transformers==4.30.2
torch==2.0.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1 